backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
requests==2.31.0
google-generativeai==0.4.0
python-multipart==0.0.9
